requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
python-dotenv==1.0.1
tweepy==4.14.0
aiohttp==3.10.10
apscheduler==3.10.4
llama-index==0.11.19
llama-index-llms-openai==0.2.16
pydantic==2.9.2
pycoingecko==3.1.0
requests==2.32.3
httpx==0.27.2
pandas==2.2.3
uvloop==0.21.0
openai==1.52.2
pytz==2024.2
pydantic-settings==2.6.0
tzlocal==5.2